+++ buidler-app/app/src/components/MyPayroll/MySalary.js
@@ -1,8 +1,8 @@
-import React from 'react'
+import React, { useMemo } from 'react'
+import { GU, useLayout } from '@aragon/ui'
 import EquityOption from '../EquityOption'
 import MySalaryChart from './MySalaryChart'
 import Salary from './Salary'
-import Split from '../Split'
 
 function MySalary() {
   return (
@@ -11,6 +11,39 @@ function MySalary() {
       <EquityOption />
       <MySalaryChart />
     </Split>
+  )
+}
+
+function Split({ children }) {
+  const { layoutName } = useLayout()
+
+  const columns = useMemo(() => {
+    if (layoutName === 'small') {
+      return '1fr'
+    }
+
+    if (layoutName === 'medium') {
+      return 'repeat(auto-fit, minmax(300px, auto))'
+    }
+
+    return '1fr 1fr 2fr'
+  }, [layoutName])
+
+  return (
+    <div
+      css={`
+        display: grid;
+        grid-template-columns: ${columns};
+        grid-gap: ${2 * GU}px;
+        margin-bottom: ${2 * GU}px;
+        ${layoutName === 'medium' &&
+          `& > :last-child {
+            grid-column: span 2;
+          }`}
+      `}
+    >
+      {children}
+    </div>
   )
 }
