--- buidler-app/app/src/components/TeamPayroll/SalaryBurnRate.js
@@ -1,9 +1,0 @@
-
-import React from 'react'
-import { Box } from '@aragon/ui'
-
-function SalaryBurnRate() {
-  return <Box heading="Salary burn rate">Salary burn rate</Box>
-}
-
-export default SalaryBurnRate
