+++ buidler-app/app/src/components/TeamPayroll/SalaryStats.js
@@ -1,16 +1,23 @@
 import React from 'react'
+import { GU, useLayout } from '@aragon/ui'
 import EquityOption from '../EquityOption'
 import PaidSalariesChart from './PaidSalariesChart'
-import SalaryBurnRate from './SalaryBurnRate'
-import Split from '../Split'
 
 function SalaryStats() {
+  const { layoutName } = useLayout()
+  const compactMode = layoutName === 'small' || layoutName === 'medium'
   return (
-    <Split>
-      <SalaryBurnRate />
-      <EquityOption readOnly={false} />
+    <div
+      css={`
+        display: grid;
+        grid-template-columns: ${compactMode ? '1fr' : '3fr 1fr'};
+        grid-gap: ${2 * GU}px;
+        margin-bottom: ${2 * GU}px;
+      `}
+    >
       <PaidSalariesChart />
-    </Split>
+      <EquityOption readOnly={false} />
+    </div>
   )
 }
